requirements.txt
boto3==1.35.87
botocore==1.35.87
jmespath==1.0.1
lxml==5.3.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-pptx==1.0.2
s3transfer==0.10.4
setuptools==75.1.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.44.0
XlsxWriter==3.2.0
